+++ src/ast.c
@@ -274,7 +274,6 @@ Text_t ast_to_sexp(ast_t *ast) {
         T(Assert, "(Assert ", ast_to_sexp(data.expr), " ", optional_sexp("message", data.message), ")");
         T(Use, "(Use ", optional_sexp("var", data.var), " ", quoted_text(data.path), ")");
         T(InlineCCode, "(InlineCCode ", ast_list_to_sexp(data.chunks), optional_type_sexp("type", data.type_ast), ")");
-        T(Deserialize, "(Deserialize ", type_ast_to_sexp(data.type), " ", ast_to_sexp(data.value), ")");
     default: errx(1, "S-expressions are not implemented for this AST");
 #undef T
     }
@@ -676,10 +675,6 @@ void ast_visit(ast_t *ast, void (*visitor)(ast_t *, void *), void *userdata) {
         ast_visit_list(Match(ast, InlineCCode)->chunks, visitor, userdata);
         return;
     }
-    case Deserialize: {
-        ast_visit(Match(ast, Deserialize)->value, visitor, userdata);
-        return;
-    }
     default: errx(1, "Visiting is not supported for this AST: %s", Text$as_c_string(ast_to_sexp(ast)));
 #undef T
     }
@@ -775,10 +770,6 @@ static void _type_ast_visit(ast_t *ast, void *userdata) {
     }
     case InlineCCode: {
         _recursive_type_ast_visit(Match(ast, InlineCCode)->type_ast, userdata);
-        break;
-    }
-    case Deserialize: {
-        _recursive_type_ast_visit(Match(ast, Deserialize)->type, userdata);
         break;
     }
     default: break;